% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/crossvalFolds.R
\name{crossvalFolds}
\alias{crossvalFolds}
\title{Creating stratified cross validation folds}
\usage{
crossvalFolds(y, k = 10, stratified = T)
}
\arguments{
\item{y}{binary vector with group assignment}

\item{k}{Default k=10. Number of folds}

\item{stratified}{Defaults to TRUE. Stratification TRUE/FALSE}
}
\description{
This function creates stratified cross validation folds, in case of unbalanced case-control groups.
}
\keyword{internal}

