% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fin_surv_model_fct.R
\name{fin_surv_model_fct}
\alias{fin_surv_model_fct}
\title{Building the final survRank model}
\usage{
fin_surv_model_fct(f, data, cv.out, nr.var = 10, c.time = NA)
}
\arguments{
\item{f}{ranking approach function. One of \code{fsSurvRankConc}, \code{fsSurvRankGlmnet}, \code{fsSurvRankRf}, \code{fsSurvRankBoost}, \code{fsSurvRankCox}, \code{fsSurvRankRandCox}, \code{fsSurvRankRpart}, \code{fsSurvRankWang}, \code{fsSurvRankitBMA} or NA, no calculation}

\item{data}{same list used as input in \code{\link{CVrankSurv_fct}}}

\item{cv.out}{number of folds in outer cross validation loop (for estimation of the predictive accuracy)}

\item{nr.var}{Number of variables up to which stepwise selection should be carried out. Has to be smaller than n number of observations.}

\item{c.time}{as defined in UnoC{survAUC} time; a positive number restricting the upper limit of the time range under consideration}
}
\value{
Output of the \code{riskscore_fct}, basically a list containing the following elements
\item{\code{ranking}}{ranking of the variables in the data set using the ranking apporach function}
\item{\code{used.rank}}{variables used in the survRank model according to the number of parameters to be used}
\item{\code{model}}{cox regression model for the selected features}
\item{\code{sum.model}}{summary of the fitted cox model}
}
\description{
This function builds the final survRank model
}
\details{
details to follow
}
\keyword{SurvRank}

