% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fixRank_fct.R
\name{fixRank_fct}
\alias{fixRank_fct}
\title{Wrapper function using f as ranking method}
\usage{
fixRank_fct(data, f, fix.var, cv.out, t.times, ncl = 2, c.time = 10, ...)
}
\arguments{
\item{...}{other arguments, not used now}

\item{input}{see details in \code{\link{riskscore_fct}}}
}
\description{
This wrapper function passes the ranking method to further functions and computes the outer survival AUCs. Setting up for parallel computing of different folds via foreach.
}
\keyword{internal}

