% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fsRankMonsterAUC_fct.R
\name{fsRankMonsterAUC_fct}
\alias{fsRankMonsterAUC_fct}
\title{Wrapper function using cox models with randomly selected nmax variables as ranking method}
\usage{
fsRankMonsterAUC_fct(data, fold, ncl, cv.out, cv.in, nr.var, t = 1,
  sd1 = 0.9, c.time, ...)
}
\arguments{
\item{...}{other arguments, not used now}

\item{input}{see details in \code{\link{CVrankSurv_fct}}}
}
\description{
This wrapper function passes the ranking method to further functions. Setting up for parallel computing of different folds via foreach.
}
\keyword{internal}

