% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fsRankRpartAUC_fct.R
\name{fsRankRpartAUC_fct}
\alias{fsRankRpartAUC_fct}
\title{Wrapper function using recursive partitioning and regression trees as ranking method}
\usage{
fsRankRpartAUC_fct(data, fold, ncl, cv.out, cv.in, nr.var, t = 1, sd1 = 0.9,
  c.time, ...)
}
\arguments{
\item{...}{other arguments, not used now}

\item{input}{see details in \code{\link{CVrankSurv_fct}}}
}
\description{
This wrapper function passes the ranking method to further functions. Setting up for parallel computing of different folds via foreach.
}
\keyword{internal}

