% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fsSurvRankBoost.R
\name{fsSurvRankBoost}
\alias{fsSurvRankBoost}
\title{Boost ranking function}
\usage{
fsSurvRankBoost(x, y, ...)
}
\arguments{
\item{x}{x matrix or data.frame}

\item{y}{response y as a survival object, generated with \code{Surv()}}

\item{...}{other arguments, not used now}
}
\description{
This function ranks the input features according to their selection probability in additive models via component-wise boosting
}
\keyword{SurvRank}

