% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fsSurvRankConc.R
\name{fsSurvRankConc}
\alias{fsSurvRankConc}
\title{Concordance ranking function}
\usage{
fsSurvRankConc(x, y, ...)
}
\arguments{
\item{x}{x matrix or data.frame}

\item{y}{response y as a survival object, generated with \code{Surv()}}

\item{...}{other arguments, not used now}
}
\description{
This function ranks the input features with the concordance measure.
}
\keyword{SurvRank}

