% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fsSurvRankCox.R
\name{fsSurvRankCox}
\alias{fsSurvRankCox}
\title{Cox ranking function}
\usage{
fsSurvRankCox(x, y, ...)
}
\arguments{
\item{x}{x matrix or data.frame}

\item{y}{response y as a survival object, generated with \code{Surv()}}

\item{...}{other arguments, not used now}
}
\description{
This function ranks the input features according to their significance in the univariate cox models
}
\keyword{SurvRank}

