% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fsSurvRankGlmnet.R
\name{fsSurvRankGlmnet}
\alias{fsSurvRankGlmnet}
\title{L1 (lasso) ranking function}
\usage{
fsSurvRankGlmnet(x, y, alp = 1, ...)
}
\arguments{
\item{x}{x matrix or data.frame}

\item{y}{response y as a survival object, generated with \code{Surv()}}

\item{alp}{alpha value in \code{glmnet} (elasticnet mixing parameter)}

\item{...}{other arguments, not used now}
}
\description{
This function ranks the input features with the lasso algorithm in glmnet.
}
\keyword{SurvRank}

