% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fsSurvRankRandCox.R
\name{fsSurvRankRandCox}
\alias{fsSurvRankRandCox}
\title{Random Cox ranking function}
\usage{
fsSurvRankRandCox(x, y, bb = 500, ...)
}
\arguments{
\item{x}{x matrix or data.frame}

\item{y}{response y as a survival object, generated with \code{Surv()}}

\item{bb}{set to 500. Number of random subsamples}

\item{...}{other arguments, not used now}
}
\description{
This function ranks the input features according to their median significance in cox models with randomly selected nmax variables
}
\keyword{SurvRank}

