% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fsSurvRankRpart.R
\name{fsSurvRankRpart}
\alias{fsSurvRankRpart}
\title{Rpart ranking function}
\usage{
fsSurvRankRpart(x, y, ...)
}
\arguments{
\item{x}{x matrix or data.frame}

\item{y}{response y as a survival object, generated with \code{Surv()}}

\item{...}{other arguments, not used now}
}
\description{
This function ranks the input features according to their importance in recursive partitioning and regression trees fitted with the function \code{rpart}
}
\keyword{SurvRank}

