% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fsSurvRankWang.R
\name{fsSurvRankWang}
\alias{fsSurvRankWang}
\title{Random Cox ranking function}
\usage{
fsSurvRankWang(x, y, bb = 400, ns = 80, ...)
}
\arguments{
\item{x}{x matrix or data.frame}

\item{y}{response y as a survival object, generated with \code{Surv()}}

\item{bb}{repeatements of the bootstrap}

\item{ns}{sample size per bootstrap}

\item{...}{other arguments, not used now}
}
\description{
This function ranks the input features according to their mean significance in univariate cox models for a randomly selected subset of the data
}
\keyword{SurvRank}

