% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmnetRank.R
\name{glmnetRank}
\alias{glmnetRank}
\title{Ranks features of a previously fitted \code{glmnet} object}
\usage{
glmnetRank(glmnet, first = T, names = T, ...)
}
\arguments{
\item{glmnet}{previously fitted \code{glmnet} object}

\item{first}{Defaults to TRUE. TRUE ranking based on occurence. FALSE based on effect sizes}

\item{names}{Defaults to TRUE. TRUE returns feature names. FALSE returns coefficients}

\item{...}{other arguments, not used now}
}
\description{
This function ranks the input features of a previously fitted \code{glmnet} object. Ranking according to the first occurence in the lambda path
or effect sizes at the end of the path.
}
\keyword{SurvRank}

