% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/msSurv_fct.R
\name{msSurv_fct}
\alias{msSurv_fct}
\title{Function selects for the ranking function approaches}
\usage{
msSurv_fct(data, fold, fs.method, ncl, cv.out, cv.in, nr.var, t = 1,
  sd1 = 0.9, c.time, ...)
}
\arguments{
\item{data}{data input}

\item{fold}{outer CV fold}

\item{fs.method}{ranking method}

\item{ncl}{number of clusters for parallel computing}

\item{cv.out}{number of folds in outer cross validation loop (for estimation of the predictive accuracy)}

\item{cv.in}{number of folds in inner cross validation loop (for model selection on the training set)}

\item{nr.var}{Number of variables up to which stepwise selection should be carried out. Has to be smaller than n number of observations.}

\item{t}{current t.times argument}

\item{c.time}{as defined in package \code{survAUC} time; a positive number restricting the upper limit of the time range under consideration.}
}
\description{
This wrapper function passes the ranking method to further functions.
}
\keyword{internal}

