% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_CVsurv.R
\name{plot_CVsurv}
\alias{plot_CVsurv}
\title{Main function of SurvRank}
\usage{
plot_CVsurv(cv.ob, data, file = "test.surv.pdf", ...)
}
\arguments{
\item{cv.ob}{output of the \code{\link{CVrankSurv_fct}}}

\item{data}{same list used as input in \code{\link{CVrankSurv_fct}}}

\item{file}{name of the pdf file to be created}

\item{...}{other arguments, not used know}
}
\description{
This function creates a pdf file containing some summary plots of the survival ranking analysis
}
\keyword{SurvRank}

