% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/weigthing_fct.R
\name{weighting_fct}
\alias{weighting_fct}
\title{Weigthing function for selected features}
\usage{
weighting_fct(out.mat, t.times, cv.out, out.sc, n1)
}
\arguments{
\item{out.mat}{matrix of dimension t.times*folds times features according to training set selections. 0/1 matrix: 1 (0) if the feature was (not) selected in the run}

\item{t.times}{number of times the cross-validation should be repeated}

\item{cv.out}{number of folds in outer cross validation loop (for estimation of the predictive accuracy)}

\item{out.sc}{matrix of t.times (of repeated cross validation) times folds (number of outer CV folds) of prediction accuracy evaluations}

\item{n1}{parameter for functions of two different weighting functions. Deviation from average performance leading to double weight of selected coefficients.}
}
\description{
Weigthing function for selected features, according to the performance in the outer loop
}
\keyword{SurvRank}

